+++ cli/service/client.go
@@ -1,17 +1,22 @@
 package service
 
 import (
+	"bytes"
 	"context"
-	"encoding/json"
 	"fmt"
 	"github.com/MinterTeam/minter-go-node/cli/pb"
 	"github.com/c-bata/go-prompt"
+	"github.com/golang/protobuf/jsonpb"
 	"github.com/golang/protobuf/proto"
+	"github.com/golang/protobuf/ptypes"
 	"github.com/golang/protobuf/ptypes/empty"
+	"github.com/marcusolsson/tui-go"
 	"github.com/urfave/cli/v2"
 	"google.golang.org/grpc"
+	"io"
 	"os"
 	"strings"
+	"time"
 )
 
 type ManagerConsole struct {
@@ -75,18 +80,23 @@ func completer(commands cli.Commands) prompt.Completer {
 	}
 }
 
-func (mc *ManagerConsole) Cli() {
+func (mc *ManagerConsole) Cli(ctx context.Context) {
 	completer := completer(mc.cli.Commands)
 	var history []string
 	for {
-		t := prompt.Input(">>> ", completer,
-			prompt.OptionHistory(history),
-			prompt.OptionShowCompletionAtStart(),
-		)
-		if err := mc.Execute(strings.Fields(t)); err != nil {
-			_, _ = fmt.Fprintln(os.Stderr, err)
+		select {
+		case <-ctx.Done():
+			return
+		default:
+			t := prompt.Input(">>> ", completer,
+				prompt.OptionHistory(history),
+				prompt.OptionShowCompletionAtStart(),
+			)
+			if err := mc.Execute(strings.Fields(t)); err != nil {
+				_, _ = fmt.Fprintln(os.Stderr, err)
+			}
+			history = append(history, t)
 		}
-		history = append(history, t)
 	}
 }
 
@@ -104,6 +114,7 @@ func ConfigureManagerConsole(socketPath string) (*ManagerConsole, error) {
 	}
 	app.UseShortOptionHandling = true
 	jsonFlag := &cli.BoolFlag{Name: "json", Aliases: []string{"j"}, Required: false, Usage: "echo in json format"}
+
 	app.Commands = []*cli.Command{
 		{
 			Name:    "dial_peer",
@@ -113,17 +124,7 @@ func ConfigureManagerConsole(socketPath string) (*ManagerConsole, error) {
 				&cli.StringFlag{Name: "address", Aliases: []string{"a"}, Required: true, Usage: "id@ip:port"},
 				&cli.BoolFlag{Name: "persistent", Aliases: []string{"p"}, Required: false},
 			},
-			Action: func(c *cli.Context) error {
-				_, err := client.DealPeer(context.Background(), &pb.DealPeerRequest{
-					Address:    c.String("address"),
-					Persistent: c.Bool("persistent"),
-				})
-				if err != nil {
-					return err
-				}
-				fmt.Println("OK")
-				return nil
-			},
+			Action: dealPeerCMD(client),
 		},
 		{
 			Name:    "prune_blocks",
@@ -133,17 +134,7 @@ func ConfigureManagerConsole(socketPath string) (*ManagerConsole, error) {
 				&cli.IntFlag{Name: "from", Aliases: []string{"f"}, Required: true},
 				&cli.IntFlag{Name: "to", Aliases: []string{"t"}, Required: true},
 			},
-			Action: func(c *cli.Context) error {
-				_, err := client.PruneBlocks(context.Background(), &pb.PruneBlocksRequest{
-					FromHeight: c.Int64("from"),
-					ToHeight:   c.Int64("to"),
-				})
-				if err != nil {
-					return err
-				}
-				fmt.Println("OK")
-				return nil
-			},
+			Action: pruneBlocksCMD(client),
 		},
 		{
 			Name:    "status",
@@ -152,22 +143,7 @@ func ConfigureManagerConsole(socketPath string) (*ManagerConsole, error) {
 			Flags: []cli.Flag{
 				jsonFlag,
 			},
-			Action: func(c *cli.Context) error {
-				response, err := client.Status(context.Background(), &empty.Empty{})
-				if err != nil {
-					return err
-				}
-				if c.Bool("json") {
-					bytes, err := json.Marshal(response)
-					if err != nil {
-						return err
-					}
-					fmt.Println(string(bytes))
-					return nil
-				}
-				fmt.Println(proto.MarshalTextString(response))
-				return nil
-			},
+			Action: statusCMD(client),
 		},
 		{
 			Name:    "net_info",
@@ -176,31 +152,19 @@ func ConfigureManagerConsole(socketPath string) (*ManagerConsole, error) {
 			Flags: []cli.Flag{
 				jsonFlag,
 			},
-			Action: func(c *cli.Context) error {
-				response, err := client.NetInfo(context.Background(), &empty.Empty{})
-				if err != nil {
-					return err
-				}
-				if c.Bool("json") {
-					bytes, err := json.Marshal(response)
-					if err != nil {
-						return err
-					}
-					fmt.Println(string(bytes))
-					return nil
-				}
-				fmt.Println(proto.MarshalTextString(response))
-				return nil
-			},
+			Action: netInfoCMD(client),
+		},
+		{
+			Name:    "dashboard",
+			Aliases: []string{"db"},
+			Usage:   "Show dashboard", //todo
+			Action:  dashboardCMD(client),
 		},
 		{
 			Name:    "exit",
 			Aliases: []string{"e"},
 			Usage:   "exit",
-			Action: func(c *cli.Context) error {
-				os.Exit(0)
-				return nil
-			},
+			Action:  exitCMD,
 		},
 	}
 
@@ -210,4 +174,216 @@ func ConfigureManagerConsole(socketPath string) (*ManagerConsole, error) {
 
 	app.Setup()
 	return NewManagerConsole(app), nil
+}
+
+func exitCMD(_ *cli.Context) error {
+	os.Exit(0)
+	return nil
+}
+
+func dashboardCMD(client pb.ManagerServiceClient) func(c *cli.Context) error {
+	return func(c *cli.Context) error {
+		ctx, cancel := context.WithCancel(c.Context)
+		response, err := client.Dashboard(ctx, &empty.Empty{})
+		if err != nil {
+			return err
+		}
+
+		defer cancel()
+
+		box := tui.NewVBox()
+		ui, err := tui.New(tui.NewHBox(box, tui.NewSpacer()))
+		if err != nil {
+			return err
+		}
+		ui.SetKeybinding("Esc", func() { ui.Quit() })
+		ui.SetKeybinding("q", func() { ui.Quit() })
+		errCh := make(chan error)
+
+		recv, err := response.Recv()
+		if err == io.EOF {
+			return nil
+		}
+		if err != nil {
+			return err
+		}
+		dashboard := updateDashboard(box, recv)
+
+		go func() { errCh <- ui.Run() }()
+
+		for {
+			select {
+			case <-c.Done():
+				return c.Err()
+			case err := <-errCh:
+				return err
+			case <-time.After(time.Second):
+				recv, err := response.Recv()
+				if err == io.EOF {
+					return nil
+				}
+				if err != nil {
+					return err
+				}
+				dashboard(recv)
+				ui.Repaint()
+			}
+		}
+	}
+}
+
+func updateDashboard(box *tui.Box, recv *pb.DashboardResponse) func(recv *pb.DashboardResponse) {
+	pubKeyText := tui.NewHBox(tui.NewLabel("Validator's Pubkey: "), tui.NewLabel(recv.ValidatorPubKey), tui.NewSpacer())
+	box.Append(pubKeyText)
+	maxProgress := pubKeyText.SizeHint().X
+	progressBox := tui.NewHBox(tui.NewEntry(), tui.NewSpacer())
+	box.Append(progressBox)
+
+	table := tui.NewTable(0, 0)
+	labelNetworkSynchronizationPercent := tui.NewLabel("")
+	labelNetworkSynchronizationTime := tui.NewLabel("")
+	table.AppendRow(tui.NewLabel("Network Synchronization"), tui.NewHBox(labelNetworkSynchronizationPercent, labelNetworkSynchronizationTime, tui.NewSpacer()))
+	labelBlockHeight := tui.NewLabel("")
+	table.AppendRow(tui.NewLabel("Block Height"), labelBlockHeight)
+	labelLastBlockTime := tui.NewLabel("")
+	table.AppendRow(tui.NewLabel("Latest Block Time"), labelLastBlockTime)
+	labelBlockProcessingTimeAvg := tui.NewLabel("")
+	table.AppendRow(tui.NewLabel("Block Processing Time (avg)"), labelBlockProcessingTimeAvg)
+	labelMemoryUsage := tui.NewLabel("")
+	table.AppendRow(tui.NewLabel("Memory Usage"), labelMemoryUsage)
+	labelPeersCount := tui.NewLabel("")
+	table.AppendRow(tui.NewLabel("Peers Count"), labelPeersCount)
+	labelValidatorStatus := tui.NewLabel("")
+	table.AppendRow(tui.NewLabel("Validator Status: "), labelValidatorStatus)
+
+	labelStakeName := tui.NewLabel("")
+	labelVotingPowerName := tui.NewLabel("")
+	labelMissedBlocksName := tui.NewLabel("")
+	labelStake := tui.NewLabel("")
+	labelVotingPower := tui.NewLabel("")
+	labelMissedBlocks := tui.NewLabel("")
+
+	table.AppendRow(labelStakeName, labelStake)
+	table.AppendRow(labelVotingPowerName, labelVotingPower)
+	table.AppendRow(labelMissedBlocksName, labelMissedBlocks)
+	box.Append(tui.NewHBox(table, tui.NewSpacer()))
+	box.Append(tui.NewSpacer())
+
+	return func(recv *pb.DashboardResponse) {
+		perSync := int((float64(recv.LatestHeight) / float64(recv.MaxPeerHeight)) * 100)
+		labelNetworkSynchronizationPercent.SetText(fmt.Sprintf("%d%% ", perSync))
+		timeLeft := ""
+		ofBlocks := ""
+		progressBox.Remove(0)
+		progressBox.Prepend(tui.NewEntry())
+		if perSync < 100 && recv.MaxPeerHeight > 0 {
+			timeLeft = "Timing..."
+			if recv.TimePerBlock != 0 {
+				timeLeft = fmt.Sprintf("(%s left)", time.Duration((recv.MaxPeerHeight-recv.LatestHeight)*recv.TimePerBlock).Truncate(time.Second).String())
+			}
+			ofBlocks = fmt.Sprintf(" of %d", recv.MaxPeerHeight)
+			progress := tui.NewProgress(maxProgress)
+			progress.SetCurrent(int(recv.LatestHeight) / (int(recv.MaxPeerHeight) / maxProgress))
+			progressBox.Remove(0)
+			progressBox.Prepend(progress)
+		}
+		labelNetworkSynchronizationTime.SetText(timeLeft)
+
+		labelBlockHeight.SetText(fmt.Sprintf("%d", recv.LatestHeight) + ofBlocks)
+		timestamp, _ := ptypes.Timestamp(recv.Timestamp)
+		labelLastBlockTime.SetText(timestamp.Format(time.RFC3339Nano) + strings.Repeat(" ", len(time.RFC3339Nano)-len(timestamp.Format(time.RFC3339Nano))))
+		labelBlockProcessingTimeAvg.SetText(fmt.Sprintf("%f sec (%f sec)", time.Duration(recv.Duration).Seconds(), time.Duration(recv.AvgBlockProcessingTime).Seconds()))
+		labelMemoryUsage.SetText(fmt.Sprintf("%d MB", recv.MemoryUsage/1024/1024))
+		labelPeersCount.SetText(fmt.Sprintf("%d", recv.PeersCount))
+		labelValidatorStatus.SetText("Not Declared")
+
+		labelStakeName.SetText("")
+		labelVotingPowerName.SetText("")
+		labelMissedBlocksName.SetText("")
+
+		labelStake.SetText("")
+		labelVotingPower.SetText("")
+		labelMissedBlocks.SetText("")
+
+		if recv.ValidatorStatus != pb.DashboardResponse_NotDeclared {
+			labelValidatorStatus.SetText(recv.ValidatorStatus.String())
+
+			labelStakeName.SetText("Stake")
+			labelVotingPowerName.SetText("Voting Power")
+			labelMissedBlocksName.SetText("Missed Blocks")
+
+			labelStake.SetText(recv.Stake)
+			labelVotingPower.SetText(fmt.Sprintf("%d", recv.VotingPower))
+			labelMissedBlocks.SetText(recv.MissedBlocks)
+		}
+	}
+}
+
+func netInfoCMD(client pb.ManagerServiceClient) func(c *cli.Context) error {
+	return func(c *cli.Context) error {
+		response, err := client.NetInfo(c.Context, &empty.Empty{})
+
+		if err != nil {
+			return err
+		}
+		if c.Bool("json") {
+			bb := new(bytes.Buffer)
+			err := (&jsonpb.Marshaler{EmitDefaults: true}).Marshal(bb, response)
+			if err != nil {
+				return err
+			}
+			fmt.Println(string(bb.Bytes()))
+			return nil
+		}
+		fmt.Println(proto.MarshalTextString(response))
+		return nil
+	}
+}
+
+func statusCMD(client pb.ManagerServiceClient) func(c *cli.Context) error {
+	return func(c *cli.Context) error {
+		response, err := client.Status(c.Context, &empty.Empty{})
+		if err != nil {
+			return err
+		}
+		if c.Bool("json") {
+			bb := new(bytes.Buffer)
+			err := (&jsonpb.Marshaler{EmitDefaults: true}).Marshal(bb, response)
+			if err != nil {
+				return err
+			}
+			fmt.Println(string(bb.Bytes()))
+			return nil
+		}
+		fmt.Println(proto.MarshalTextString(response))
+		return nil
+	}
+}
+
+func pruneBlocksCMD(client pb.ManagerServiceClient) func(c *cli.Context) error {
+	return func(c *cli.Context) error {
+		_, err := client.PruneBlocks(c.Context, &pb.PruneBlocksRequest{
+			FromHeight: c.Int64("from"),
+			ToHeight:   c.Int64("to"),
+		})
+		if err != nil {
+			return err
+		}
+		fmt.Println("OK")
+		return nil
+	}
+}
+
+func dealPeerCMD(client pb.ManagerServiceClient) func(c *cli.Context) error {
+	return func(c *cli.Context) error {
+		_, err := client.DealPeer(c.Context, &pb.DealPeerRequest{
+			Address:    c.String("address"),
+			Persistent: c.Bool("persistent"),
+		})
+		if err != nil {
+			return err
+		}
+		fmt.Println("OK")
+		return nil
+	}
 }