+++ core/statistics/statistics.go
@@ -2,87 +2,126 @@ package statistics
 
 import (
 	"context"
-	"crypto/tls"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/tendermint/tendermint/rpc/core"
 	rpctypes "github.com/tendermint/tendermint/rpc/lib/types"
-	"net/http"
-	"net/http/httptrace"
+	"net"
 	"net/url"
+	"runtime"
 	"sync"
 	"time"
 )
 
+type ping struct {
+	duration time.Duration
+	url      string
+}
+
 func (d *Data) Statistic(ctx context.Context) {
-	http.DefaultTransport.(*http.Transport).TLSClientConfig = &tls.Config{InsecureSkipVerify: true}
+	go d.handleStartBlocks(ctx)
+	go d.handleEndBlock(ctx)
 	for {
 		select {
 		case <-ctx.Done():
 			return
-		case <-time.After(time.Second * 10): //todo embedded period
+		case <-time.After(time.Second * 10):
 			state, err := core.NetInfo(&rpctypes.Context{})
 			if err != nil {
 				continue
 			}
-
 			var wg sync.WaitGroup
-			wg.Add(len(state.Peers))
-			d.Peer.Lock()
+			wg.Add(state.NPeers)
+			c := make(chan *ping, state.NPeers)
 			for _, peer := range state.Peers {
-				u := &url.URL{Scheme: "http", Host: peer.RemoteIP}
-				go func() {
+				parse, err := url.Parse(peer.NodeInfo.ListenAddr)
+				if err != nil {
+					continue
+				}
+				go func(s string) {
 					defer wg.Done()
-					s := u.String()
-					duration, err := timeGet(s)
+					duration, err := pingTCP(s)
 					if err != nil {
 						return
 					}
-					d.SetPeerTime(duration, s)
-				}()
+					c <- &ping{
+						duration: duration,
+						url:      s,
+					}
+				}(net.JoinHostPort(peer.RemoteIP, parse.Port()))
 			}
-			d.Peer.ping.Reset()
-			d.Peer.Unlock()
 			wg.Wait()
+			d.ResetPeersPing()
+			close(c)
+			for ping := range c {
+				d.SetPeerTime(ping.duration, ping.url)
+			}
 		}
 	}
 }
 
-func timeGet(url string) (time.Duration, error) {
-	req, err := http.NewRequest("GET", url, nil)
-	if err != nil {
-		return 0, err
-	}
-	req = req.WithContext(httptrace.WithClientTrace(req.Context(), &httptrace.ClientTrace{}))
+func pingTCP(url string) (time.Duration, error) {
 	start := time.Now()
-	_, err = http.DefaultTransport.RoundTrip(req)
+	conn, err := net.DialTimeout("tcp", url, 5*time.Second)
 	if err != nil {
 		return 0, err
 	}
+	defer conn.Close()
 	return time.Since(start), nil
 }
 
 type Data struct {
 	BlockStart struct {
-		sync.Mutex
-		height    uint64
-		time      time.Time
-		timestamp float64
+		sync.RWMutex
+		height          int64
+		time            time.Time
+		headerTimestamp time.Time
 	}
 	BlockEnd blockEnd
+	cS       chan StartRequest
+	cE       chan EndRequest
+	Speed    struct {
+		sync.RWMutex
+		startTime         time.Time
+		startHeight       int64
+		duration          int64
+		timerMin          <-chan time.Time
+		blocksCountPerMin int64
+		avgTimePerBlock   int64
+	}
 
 	Api  apiResponseTime
 	Peer peerPing
 }
-type blockEnd struct {
-	sync.Mutex
-	Height    prometheus.Gauge
-	Duration  prometheus.Gauge
-	Timestamp prometheus.Gauge
+
+type StartRequest struct {
+	Height     int64
+	Now        time.Time
+	HeaderTime time.Time
 }
+type EndRequest struct {
+	TimeEnd time.Time
+	Height  int64
+}
+
+type LastBlockInfo struct {
+	Height          int64
+	Duration        int64
+	HeaderTimestamp time.Time
+}
+
+type blockEnd struct {
+	sync.RWMutex
+	HeightProm    prometheus.Gauge
+	DurationProm  prometheus.Gauge
+	TimestampProm prometheus.Gauge
+	LastBlockInfo LastBlockInfo
+}
+
 type apiResponseTime struct {
 	sync.Mutex
 	responseTime *prometheus.GaugeVec
 }
+
 type peerPing struct {
 	sync.Mutex
 	ping *prometheus.GaugeVec
@@ -92,7 +131,7 @@ func New() *Data {
 	apiVec := prometheus.NewGaugeVec(
 		prometheus.GaugeOpts{
 			Name: "api",
-			Help: "Api Duration Paths",
+			Help: "Api DurationProm Paths",
 		},
 		[]string{"path"},
 	)
@@ -122,7 +161,7 @@ func New() *Data {
 	timeBlock := prometheus.NewGauge(
 		prometheus.GaugeOpts{
 			Name: "last_block_timestamp",
-			Help: "Timestamp last block",
+			Help: "TimestampProm last block",
 		},
 	)
 	prometheus.MustRegister(timeBlock)
@@ -130,41 +169,137 @@ func New() *Data {
 	return &Data{
 		Api:      apiResponseTime{responseTime: apiVec},
 		Peer:     peerPing{ping: peerVec},
-		BlockEnd: blockEnd{Height: height, Duration: lastBlockDuration, Timestamp: timeBlock},
+		BlockEnd: blockEnd{HeightProm: height, DurationProm: lastBlockDuration, TimestampProm: timeBlock},
+		cS:       make(chan StartRequest),
+		cE:       make(chan EndRequest),
 	}
 }
 
-func (d *Data) SetStartBlock(height uint64, now time.Time, headerTime time.Time) {
+func (d *Data) PushStartBlock(req StartRequest) {
+	if d == nil {
+		return
+	}
+	d.cS <- req
+}
+
+func (d *Data) handleStartBlocks(ctx context.Context) {
 	if d == nil {
 		return
 	}
 
-	d.BlockStart.Lock()
-	defer d.BlockStart.Unlock()
+	for {
+		select {
+		case <-ctx.Done():
+			return
+		case req := <-d.cS:
+			func() {
+				height := req.Height
+				now := req.Now
+				headerTime := req.HeaderTime
 
-	d.BlockStart.height = height
-	d.BlockStart.time = now
-	d.BlockStart.timestamp = float64(headerTime.UnixNano() / 1e09)
+				for {
+					d.BlockStart.RLock()
+					ok := (height == d.BlockStart.height+1) || 0 == d.BlockStart.height
+					d.BlockStart.RUnlock()
+					if ok {
+						break
+					}
+					runtime.Gosched()
+				}
+
+				d.BlockStart.Lock()
+				defer d.BlockStart.Unlock()
+
+				d.BlockStart.height = height
+				d.BlockStart.time = now
+				d.BlockStart.headerTimestamp = headerTime
+			}()
+		}
+	}
 }
 
-func (d *Data) SetEndBlockDuration(timeEnd time.Time, height uint64) {
+func (d *Data) PushEndBlock(req EndRequest) {
 	if d == nil {
 		return
 	}
+	d.cE <- req
+}
 
-	d.BlockStart.Lock()
-	defer d.BlockStart.Unlock()
-
-	if height == d.BlockStart.height {
-		d.BlockEnd.Lock()
-		defer d.BlockEnd.Unlock()
-		d.BlockEnd.Height.Set(float64(height))
-		d.BlockEnd.Duration.Set(timeEnd.Sub(d.BlockStart.time).Seconds())
-		d.BlockEnd.Timestamp.Set(d.BlockStart.timestamp)
+func (d *Data) handleEndBlock(ctx context.Context) {
+	if d == nil {
 		return
 	}
+	for {
+		select {
+		case <-ctx.Done():
+			return
+		case req := <-d.cE:
+			func() {
+				height := req.Height
+				timeEnd := req.TimeEnd
 
-	return
+				for {
+					d.BlockStart.RLock()
+					d.BlockEnd.RLock()
+					ok := height == d.BlockStart.height
+					d.BlockEnd.RUnlock()
+					d.BlockStart.RUnlock()
+					if ok {
+						break
+					}
+					runtime.Gosched()
+				}
+
+				d.BlockStart.Lock()
+				defer d.BlockStart.Unlock()
+
+				duration := timeEnd.Sub(d.BlockStart.time)
+
+				d.BlockEnd.Lock()
+				defer d.BlockEnd.Unlock()
+
+				d.BlockEnd.HeightProm.Set(float64(height))
+				d.BlockEnd.DurationProm.Set(duration.Seconds())
+				d.BlockEnd.TimestampProm.Set(float64(d.BlockStart.headerTimestamp.UnixNano()))
+
+				d.BlockEnd.LastBlockInfo.Height = height
+				d.BlockEnd.LastBlockInfo.Duration = duration.Nanoseconds()
+				d.BlockEnd.LastBlockInfo.HeaderTimestamp = d.BlockStart.headerTimestamp
+
+				d.Speed.Lock()
+				defer d.Speed.Unlock()
+
+				min := time.Minute
+				select {
+				case <-d.Speed.timerMin:
+					d.Speed.avgTimePerBlock = int64(min) / d.Speed.blocksCountPerMin
+					d.Speed.timerMin = time.After(min)
+					d.Speed.blocksCountPerMin = 1
+				default:
+					d.Speed.blocksCountPerMin++
+				}
+
+				if time.Since(d.Speed.startTime) < 24*time.Hour {
+					d.Speed.duration += duration.Nanoseconds()
+					return
+				}
+
+				if d.Speed.startHeight == 0 {
+					d.Speed.startTime = time.Now()
+					d.Speed.startHeight = height
+					d.Speed.duration = duration.Nanoseconds()
+					d.Speed.timerMin = time.After(min)
+					return
+				}
+
+				d.Speed.startTime = time.Now().Add(-12 * time.Hour)
+				d.Speed.startHeight = height - (height-d.Speed.startHeight)/2
+				d.Speed.duration = d.Speed.duration/2 + duration.Nanoseconds()
+
+				return
+			}()
+		}
+	}
 }
 
 func (d *Data) SetApiTime(duration time.Duration, path string) {
@@ -187,4 +322,51 @@ func (d *Data) SetPeerTime(duration time.Duration, network string) {
 	defer d.Peer.Unlock()
 
 	d.Peer.ping.With(prometheus.Labels{"network": network}).Set(duration.Seconds())
+}
+
+func (d *Data) GetAverageBlockProcessingTime() int64 {
+	if d == nil {
+		return 0
+	}
+
+	d.Speed.RLock()
+	defer d.Speed.RUnlock()
+
+	d.BlockEnd.RLock()
+	defer d.BlockEnd.RUnlock()
+
+	return d.Speed.duration / (d.BlockEnd.LastBlockInfo.Height - d.Speed.startHeight)
+}
+
+func (d *Data) GetTimePerBlock() int64 {
+	if d == nil {
+		return 0
+	}
+
+	d.Speed.RLock()
+	defer d.Speed.RUnlock()
+
+	return d.Speed.avgTimePerBlock
+}
+
+func (d *Data) GetLastBlockInfo() LastBlockInfo {
+	if d == nil {
+		return LastBlockInfo{}
+	}
+
+	d.BlockEnd.RLock()
+	defer d.BlockEnd.RUnlock()
+
+	return d.BlockEnd.LastBlockInfo
+}
+
+func (d *Data) ResetPeersPing() {
+	if d == nil {
+		return
+	}
+
+	d.Peer.Lock()
+	defer d.Peer.Unlock()
+
+	d.Peer.ping.Reset()
 }