+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 1.1.6
+
+IMPROVEMENT
+
+- [export] Added export command
+- [db] Update IAVL to v0.13.2
+- [console] Added dashboard command
+- [docker] Fix docker build config (@dmitry-ee)
+
+BUG FIXES
+
+- [prometheus] Fix too many open descriptors problem
+
 ## 1.1.5
 
 IMPROVEMENT